Moved the CSRFToken to the shared folder static/src/shared/csrf.js
