+++ plannr/static/src/app.js
@@ -19,4 +19,4 @@ const App = () => (
     </div>
   </Router>
 )
-export default App 
+export default App
