+++ plannr/templates/global/base.html
@@ -25,6 +25,7 @@
                 <li class="active"><a href="month">Month</a></li>
                 <li><a href="week">Week</a></li>
                 <li><a href="#contact">Day</a></li>
+                <div id="logout"></div>
               </ul>
             </div>
           </div>
